Moved cpu_usage_sampler to system_info, renamed where needed to system_cpu_sampler, fixed clang format
