+++ source/darwin/system_cpu_usage_sampler.c
